requirements.txt
click>=8.0
rich>=13.0
requests>=2.25
pydantic>=2.0
typer>=0.9
fastapi>=0.104.0
uvicorn>=0.24.0
pillow>=9.0
packaging>=21.0
posthog>=3.0.0